Added calloc and realloc functions

+++ include/malloc.h
@@ -1,7 +1,9 @@
 #ifndef _MALLOC_H_
 #define _MALLOC_H_
 
-void * malloc(int n);
+#include <stddef.h>
+
+void * malloc(size_t n);
 void free(void *p);
 
 #endif // _MALLOC_H_
