+++ src/string/memcpy.c
@@ -3,7 +3,7 @@
 
 #include <string.h>
 
-void * memcpy (void *dest, const void *src, int len) {
+void * memcpy (void *dest, const void *src, size_t len) {
   char *d = dest;
   const char *s = src;
   while (len--) {
